修改服务注册发现包位置
