+++ app/shop/admin/internal/server/server.go
@@ -1,8 +1,36 @@
 package server
 
 import (
+	"github.com/go-kratos/beer-shop/app/shop/admin/internal/conf"
+	consul "github.com/go-kratos/consul/registry"
+	"github.com/go-kratos/kratos/v2/registry"
 	"github.com/google/wire"
+	consulAPI "github.com/hashicorp/consul/api"
 )
 
 // ProviderSet is server providers.
-var ProviderSet = wire.NewSet(NewHTTPServer, NewGRPCServer)
+var ProviderSet = wire.NewSet(NewHTTPServer, NewGRPCServer, NewDiscovery, NewRegistrar)
+
+func NewDiscovery(conf *conf.Registry) registry.Discovery {
+	c := consulAPI.DefaultConfig()
+	c.Address = conf.Consul.Address
+	c.Scheme = conf.Consul.Scheme
+	cli, err := consulAPI.NewClient(c)
+	if err != nil {
+		panic(err)
+	}
+	r := consul.New(cli, consul.WithHealthCheck(false))
+	return r
+}
+
+func NewRegistrar(conf *conf.Registry) registry.Registrar {
+	c := consulAPI.DefaultConfig()
+	c.Address = conf.Consul.Address
+	c.Scheme = conf.Consul.Scheme
+	cli, err := consulAPI.NewClient(c)
+	if err != nil {
+		panic(err)
+	}
+	r := consul.New(cli, consul.WithHealthCheck(false))
+	return r
+}
